requirements.txt
django
djangorestframework
django-filter
psycopg2-binary
python-dotenv
djangorestframework-simplejwt
django-cors-headers
